sdk extraction > cleanup
