--- src/main/java/com/hivemq/http/core/MimeTypeNotFoundException.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2019-present HiveMQ GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hivemq.http.core;
-
-import java.io.IOException;
-
-public class MimeTypeNotFoundException extends IOException{
-
-    public MimeTypeNotFoundException() {
-    }
-
-    public MimeTypeNotFoundException(String message) {
-        super(message);
-    }
-
-    public MimeTypeNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public MimeTypeNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
