+++ src/main/java/com/hivemq/mqtt/PublishReturnCode.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hivemq.mqtt.handler.publish;
+package com.hivemq.mqtt;
 
 import com.hivemq.extension.sdk.api.annotations.NotNull;
